# perf: reduce std::string_view creation