Before

```
BasicBench_AdaURL       4825 ns         4824 ns       106667 time/byte=5.97064ns time/url=438.571ns url/s=2.28014M/s
```

After

```
BasicBench_AdaURL            4478 ns         4477 ns       156362 time/byte=5.54135ns time/url=407.037ns url/s=2.45678M/s
```